패키지, 컨트롤러 이름 수정
