+++ src/main/java/com/oti/team2/jobgrade/dao/IJobGradeDao.java
@@ -1,8 +1,8 @@
-package com.oti.team2.jobGrade.dao;
+package com.oti.team2.jobgrade.dao;
 
 import java.util.List;
 
-import com.oti.team2.jobGrade.dto.JobGrade;
+import com.oti.team2.jobgrade.dto.JobGrade;
 /*
  * @author: 안한길
  * 작성일: 23.02.21.
