+++ src/main/java/com/oti/team2/jobgrade/service/JobGradeService.java
@@ -1,12 +1,12 @@
-package com.oti.team2.jobGrade.service;
+package com.oti.team2.jobgrade.service;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.oti.team2.jobGrade.dao.IJobGradeDao;
-import com.oti.team2.jobGrade.dto.JobGrade;
+import com.oti.team2.jobgrade.dao.IJobGradeDao;
+import com.oti.team2.jobgrade.dto.JobGrade;
 /* 직급 서비스
  * @author: 안한길
  * */
