+++ src/main/resources/mybatis/mapper/mapperJobGrade.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.oti.team2.jobGrade.dao.IJobGradeDao">
+<mapper namespace="com.oti.team2.jobgrade.dao.IJobGradeDao">
 	<!-- 직급 리스트 -->
 	<select id="selectJobGradeList" resultType="jobGrade">
 		SELECT
