+++ src/main/resources/spring/mybatis.xml
@@ -16,7 +16,7 @@
 	</bean>
 
 
-	<mybatis-spring:scan base-package="com.oti.team2.jobGrade.dao"/>
+	<mybatis-spring:scan base-package="com.oti.team2.jobgrade.dao"/>
 	<mybatis-spring:scan base-package="com.oti.team2.member.dao"/>
 	<mybatis-spring:scan base-package="com.oti.team2.department.dao"/>
 	<mybatis-spring:scan base-package="com.oti.team2.institution.dao"/>
